Cleanup. Fix tool description parsing
