+++ plugins/agents/agent-toolkit/src/hopeit_agents/agent_toolkit/agents/__init__.py
@@ -1,9 +1,1 @@
 """Agent-specific helpers exposed by the agent toolkit package."""
-
-from .prompts import AgentConfig, create_agent_config, render_prompt
-
-__all__ = [
-    "AgentConfig",
-    "create_agent_config",
-    "render_prompt",
-]
