+++ plugins/agents/agent-toolkit/src/hopeit_agents/agent_toolkit/mcp/__init__.py
@@ -1,13 +1,1 @@
 """Utilities that integrate the agent toolkit with the MCP client."""
-
-from .agent_tools import (
-    call_tool,
-    execute_tool_calls,
-    tool_descriptions,
-)
-
-__all__ = [
-    "call_tool",
-    "execute_tool_calls",
-    "tool_descriptions",
-]
