+++ plugins/agents/model-client/src/hopeit_agents/model_client/api/__init__.py
@@ -1,1 +1,1 @@
-"""Public API exports for the `hopeit_agents.model_client` package."""
+"""hopeit endpoints for the `hopeit_agents.model_client` plugin."""
