+++ plugins/mcp/mcp-client/src/hopeit_agents/mcp_client/api/__init__.py
@@ -1,1 +1,1 @@
-"""Expose the public MCP client API surface."""
+"""Expose the public MCP client as hopeit endpoints surface."""
