+++ plugins/mcp/mcp-server/src/hopeit_agents/mcp_server/server/__init__.py
@@ -1,3 +1,3 @@
 """
-hopeit-agents-mcp-server CLI (Command Line Interface) module
+hopeit-agents-mcp-server server module
 """
